feat(mcp): implement tool registry abstraction and code generation for MCP tools
