+++ src/modules/mcp/mcp.tools.ts
@@ -1,4 +1,4 @@
-import type { AnyToolDef } from '@/modules/mcp/tool-registry';
+import type { AnyToolDef } from '@/modules/mcp/types';
 
 // Placeholder — module tool files (e.g. src/modules/matters/mcp.tools.ts) will be added here
 // as they are implemented and auto-discovered via codegen in a follow-up PR.
