+++ src/modules/mcp/server.ts
@@ -1,13 +1,13 @@
 import { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
 import { WebStandardStreamableHTTPServerTransport } from '@modelcontextprotocol/sdk/server/webStandardStreamableHttp.js';
 
-import { registerTools } from '@/modules/mcp/tool-registry';
+import { toolRegistry } from '@/modules/mcp/tool-registry';
 import { allTools } from '@/modules/mcp/mcp.tools';
 import type { McpJwt } from '@/modules/mcp/types';
 
 export const createMcpServer = (jwt: McpJwt): McpServer => {
   const server = new McpServer({ name: 'blawby', version: '1.0.0' });
-  registerTools(server, jwt, allTools);
+  toolRegistry.registerTools(server, jwt, allTools);
   return server;
 };
